// helpers/action-types.ts
export enum DATA_ACTION_TYPES {
    SET_LOCATION = 'SET_LOCATION',
    SET_CHECK_IN = 'SET_CHECK_IN',
    SET_CHECK_OUT = 'SET_CHECK_OUT',
    SET_GUESTS = 'SET_GUESTS',
    RESET_DATES = 'RESET_DATES',
    RESET_GUESTS = 'RESET_GUESTS',
    INCREASE_ADULTS = 'INCREASE_ADULTS',
    INCREASE_CHILDREN = 'INCREASE_CHILDREN',
    INCREASE_INFANTS = 'INCREASE_INFANTS',
    INCREASE_PETS = 'INCREASE_PETS',
    DECREASE_ADULTS = 'DECREASE_ADULTS',
    DECREASE_CHILDREN = 'DECREASE_CHILDREN',
    DECREASE_INFANTS = 'DECREASE_INFANTS',
    DECREASE_PETS = 'DECREASE_PETS',
  }
  